+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - fixed issue with rating TV Series/Mini Series
 - switched to media-imdb.com for fetching movie titles
 - fixed: wrong title (ie: "Primetime Emmys ()") saved when IMDb puts a promoted entry first in the suggestions API results
+- TV Series titles (but not TV Mini Series) now include a season marker, ie: "Your Friends & Neighbors S01 (TV Series, 2025)"
 
 ## Release 1.9 (2025-11-16)